moved color picker to components
