+++ Assets/UI Toolkit/Resources/UI/Components/ColorPicker-style.uss
