+++ Assets/UI Toolkit/Scripts/IPropertyPanelWithColorPicker.cs
@@ -1,3 +1,5 @@
+using Netherlands3D.UI.Components;
+
 namespace Netherlands3D.UI.Panels
 {
     public interface IPropertyPanelWithColorPicker
